+++ Agraria.Repositorio/Repositorios/StockRepository.cs
@@ -88,9 +88,9 @@ namespace Agraria.Repositorio.Repositorios
                     Stock stock = new()
                     {
                         Cod_Articulo = reader.GetString(0),
-                        Cantidad = reader.GetDouble(1),
-                        Costo = reader.GetDouble(2),
-                        Ganancia = reader.GetDouble(3)
+                        Cantidad = reader.GetDecimal(1),
+                        Costo = reader.GetDecimal(2),
+                        Ganancia = reader.GetDecimal(3)
 
                     };
                     stocks.Add(stock);
@@ -121,9 +121,9 @@ namespace Agraria.Repositorio.Repositorios
                     Stock stock = new()
                     {
                         Cod_Articulo = reader.GetString(0),
-                        Cantidad = reader.GetDouble(1),
-                        Costo = reader.GetDouble(2),
-                        Ganancia = reader.GetDouble(3)
+                        Cantidad = reader.GetDecimal(1),
+                        Costo = reader.GetDecimal(2),
+                        Ganancia = reader.GetDecimal(3)
                     };
                     return Result<Stock>.Success(stock);
                 }
